# Improve the performance of Adler-32 and decoding. Validate against .NET 5