--- a/.github/workflows/build-test.yml
+++ b/.github/workflows/build-test.yml
@@ -1,4 +1,4 @@
-name: .NET Core
+name: Build and test
 
 on: [push, pull_request]
 
@@ -12,11 +12,19 @@ jobs:
 
     steps:
     - uses: actions/checkout@v1
-    - name: Setup .NET SDK
+
+    - name: "Setup .NET SDK"
       uses: actions/setup-dotnet@v1
       with:
-        dotnet-version: 3.1.100
+        dotnet-version: 5.0.101
+
+    - name: Setup .NET 3.1 runtime
+      uses: actions/setup-dotnet@v1
+      with:
+        dotnet-version: 3.1.404
+
     - name: Build
       run: dotnet build -c Release
+
     - name: Test
       run: dotnet test -c Release --no-build