--- a/.gitignore
+++ b/.gitignore
@@ -9,5 +9,8 @@ obj/
 # NuGet
 packages/
 
-# NUnit output
+# Test output
 TestResult.xml
+artifacts/
+coverage/
+tools/