--- a/.gitignore
+++ b/.gitignore
@@ -14,3 +14,4 @@ TestResult.xml
 artifacts/
 coverage/
 tools/
+xdelta.PerformanceTests/BenchmarkDotNet.Artifacts/*