--- a/xdelta.Cli/xdelta.Cli.csproj
+++ b/xdelta.Cli/xdelta.Cli.csproj
@@ -8,7 +8,7 @@
     <Authors>Benito Palacios Sánchez</Authors>
     <Copyright>Copyright (C) 2015 Benito Palacios Sánchez</Copyright>
 
-    <Version>1.1</Version>
+    <Version>1.2</Version>
     <TargetFramework>netcoreapp3.1</TargetFramework>
     <GenerateDocumentationFile>false</GenerateDocumentationFile>
     <CodeAnalysisRuleSet></CodeAnalysisRuleSet>
@@ -17,7 +17,7 @@
   <PropertyGroup>
     <PackageProjectUrl>https://github.com/pleonex/xdelta-sharp</PackageProjectUrl>
     <RepositoryUrl>https://github.com/pleonex/xdelta-sharp</RepositoryUrl>
-    <License>GPL-3</License>
+    <License>MIT</License>
     <PackageReleaseNotes></PackageReleaseNotes>
     <PackageIcon></PackageIcon>
     <PackageTags>decompress;vcdiff;rfc3284;xdelta</PackageTags>