--- a/xdelta.UnitTests/xdelta.UnitTests.csproj
+++ b/xdelta.UnitTests/xdelta.UnitTests.csproj
@@ -10,6 +10,7 @@
 
     <Version>1.1</Version>
     <TargetFrameworks>netcoreapp3.1;net48;net5.0</TargetFrameworks>
+    <LangVersion>8.0</LangVersion>
     <GenerateDocumentationFile>false</GenerateDocumentationFile>
   </PropertyGroup>
 
@@ -19,6 +20,7 @@
     <PackageReference Include="NUnit" Version="3.12.0" />
     <PackageReference Include="NUnit3TestAdapter" Version="3.15.1" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.4.0" />
+    <PackageReference Include="coverlet.collector" Version="1.3.0" />
 
     <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.0">
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>