--- a/xdelta.sln
+++ b/xdelta.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "xdelta.UnitTests", "xdelta.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "xdelta.Cli", "xdelta.Cli\xdelta.Cli.csproj", "{773349A8-4924-4C83-857C-522EF949B813}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Xdelta.PerformanceTests", "xdelta.PerformanceTests\Xdelta.PerformanceTests.csproj", "{DF38AA30-930F-41D4-8555-D85689AEE32B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -58,5 +60,17 @@ Global
 		{773349A8-4924-4C83-857C-522EF949B813}.Release|x64.Build.0 = Release|Any CPU
 		{773349A8-4924-4C83-857C-522EF949B813}.Release|x86.ActiveCfg = Release|Any CPU
 		{773349A8-4924-4C83-857C-522EF949B813}.Release|x86.Build.0 = Release|Any CPU
+		{DF38AA30-930F-41D4-8555-D85689AEE32B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DF38AA30-930F-41D4-8555-D85689AEE32B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DF38AA30-930F-41D4-8555-D85689AEE32B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{DF38AA30-930F-41D4-8555-D85689AEE32B}.Debug|x64.Build.0 = Debug|Any CPU
+		{DF38AA30-930F-41D4-8555-D85689AEE32B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{DF38AA30-930F-41D4-8555-D85689AEE32B}.Debug|x86.Build.0 = Debug|Any CPU
+		{DF38AA30-930F-41D4-8555-D85689AEE32B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DF38AA30-930F-41D4-8555-D85689AEE32B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DF38AA30-930F-41D4-8555-D85689AEE32B}.Release|x64.ActiveCfg = Release|Any CPU
+		{DF38AA30-930F-41D4-8555-D85689AEE32B}.Release|x64.Build.0 = Release|Any CPU
+		{DF38AA30-930F-41D4-8555-D85689AEE32B}.Release|x86.ActiveCfg = Release|Any CPU
+		{DF38AA30-930F-41D4-8555-D85689AEE32B}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal