--- a/xdelta/xdelta.csproj
+++ b/xdelta/xdelta.csproj
@@ -8,7 +8,7 @@
     <Authors>Benito Palacios Sánchez</Authors>
     <Copyright>Copyright (C) 2015 Benito Palacios Sánchez</Copyright>
 
-    <Version>1.1</Version>
+    <Version>1.2</Version>
     <TargetFrameworks>netstandard2.0;net461</TargetFrameworks>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <CodeAnalysisRuleSet></CodeAnalysisRuleSet>