Decoding a window of data requires running the checksum algorithm Adler-32 to verify the decoding was successful. There was a major performance issue and this checksum algorithm was very complex and inefficient.

I have added unit tests and a performance test. Results:

Original

```
|     Method |  Length |          Mean |       Error |      StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |-------- |--------------:|------------:|------------:|------:|------:|------:|----------:|
| Adler32Run |    2048 |      5.999 us |   0.1135 us |   0.1627 us |     - |     - |     - |         - |
| Adler32Run |  524288 |  1,583.961 us |  29.7528 us |  56.6077 us |     - |     - |     - |       1 B |
| Adler32Run | 4194304 | 13,057.855 us | 247.3469 us | 553.2267 us |     - |     - |     - |       4 B |
```

After improvement

```
|     Method |  Length |         Mean |      Error |     StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------- |-------- |-------------:|-----------:|-----------:|------:|------:|------:|----------:|
| Adler32Run |    2048 |     1.357 us |  0.0111 us |  0.0093 us |     - |     - |     - |         - |
| Adler32Run |  524288 |   333.501 us |  1.7935 us |  1.5899 us |     - |     - |     - |         - |
| Adler32Run | 4194304 | 2,760.907 us | 15.1725 us | 12.6697 us |     - |     - |     - |       1 B |
```

2 KB   -> 77.3%
512 KB -> 78.9%
4 MB   -> 78.9%

Also validate the tests against .NET 5